creates new copy to clipboard component
