+++ app/components/shared/index.js
@@ -1,1 +1,3 @@
 export { default as Tooltip } from "./Tooltip";
+
+export const Aux = ({ children }) => children;
