+++ app/components/views/TransactionsPage/ReceiveTab/Page.js
@@ -4,7 +4,7 @@ import ReactTooltip from "react-tooltip";
 import ReceiveAccountsSelect from "../../../ReceiveAccountsSelect";
 import { Link } from "react-router";
 import KeyBlueButton from "../../../KeyBlueButton";
-import CopyToClipboardButton from "../../../CopyToClipboardButton";
+import { CopyToClipboard } from "shared";
 import QRCode from "./QRCode";
 import { defineMessages, FormattedMessage as T, injectIntl } from "react-intl";
 import "../../../../style/Layout.less";
@@ -45,11 +45,7 @@ const ReceivePage = ({
       <div className="receive-content-nest-qr">
         <div className="receive-content-nest-qrhash">
           <span key="addressSpan">{nextAddress}</span>
-          <CopyToClipboardButton
-            key="copyToClipboard"
-            className="receive-content-nest-copy-to-clipboard-icon"
-            textToCopy={nextAddress}
-          />
+          <CopyToClipboard textToCopy={nextAddress} className="receive-content-nest-copy-to-clipboard-icon" />
         </div>
         <QRCode addr={nextAddress} />
         <div style={{ clear: "both" }}></div>
